create separate .png files from the base64 repr for ease of review
